修复最高音频无法使用问题

+++ include/config.h
@@ -58,7 +58,7 @@ namespace config
     {
       bool start = false;
       uint8_t channel = 2;
-      uint16_t rate = 48000;
+      uint32_t rate = 48000;
       uint8_t bit = 16;
       bool autoVolumn = false;
       bool peekVolumn = false;
